+++ js/pkg/garu_wasm.js
@@ -121,7 +121,7 @@ export class GaruWasm {
     }
 }
 if (Symbol.dispose) GaruWasm.prototype[Symbol.dispose] = GaruWasm.prototype.free;
-import * as import1 from "./snippets/garu-core-36fcadacb06b8856/inline0.js"
+import * as import1 from "./snippets/garu-core-996c7dd407e72272/inline0.js"
 
 function __wbg_get_imports() {
     const import0 = {
@@ -180,7 +180,7 @@ function __wbg_get_imports() {
     return {
         __proto__: null,
         "./garu_wasm_bg.js": import0,
-        "./snippets/garu-core-36fcadacb06b8856/inline0.js": import1,
+        "./snippets/garu-core-996c7dd407e72272/inline0.js": import1,
     };
 }